+++ packages/test-project/vue-adapter-options.js
@@ -1,5 +1,0 @@
-const { resolve } = require('path')
-
-module.exports = {
-  bundle: resolve(__dirname, '../../test/fixtures/vue-server.js')
-}